Switch lines to using rectangles
This means that the lines will look smaller the further way the camera
is from the line. Just increasing the line size means that all lines are
the same size, regardless of distance. Now, rectangle billboards that
always face towards the camera along the direction perpendicular to the
line are used.