Use UBO in text shaders and tidy engine loop
I need to improve the UBO and work out the padding as I dont like having
cameraPosition as a vec4 when it can be a vec3. Also, I need to figure
out the best way or storing the matrices in the camera class. Do I use a
struct to store the main variables and I can return those all at once?